% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{bb_person_score}
\alias{bb_person_score}
\title{Calculate observed score for each participant}
\usage{
bb_person_score(fbi_bb_data, scored_responses)
}
\arguments{
\item{fbi_bb_data}{Dataset of the form of TestResponses}

\item{scored_responses}{Scored responses (result of score_bb_data())}
}
\value{
Data frame of scores, one row for each participant
}
\description{
Calculate observed score for each participant
}
\examples{
bb_person_score(TestResponses, im_scored)
}
