% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{error_rate_analysis}
\alias{error_rate_analysis}
\title{Calculate error rates and avg question difficulty for each participant}
\usage{
error_rate_analysis(fbi_bb_data, q_diff)
}
\arguments{
\item{fbi_bb_data}{Dataset of the form of TestResponses}

\item{q_diff}{Question difficulty (from IRT analysis)}
}
\value{
Data frame of avg difficulty, percent of items skipped, FPR, FNR, one row for each participant
}
\description{
Calculate error rates and avg question difficulty for each participant
}
\examples{
\dontrun{
error_rate_analysis(TestResponses, question_difficulties)
}
}
