% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{fit_irt}
\alias{fit_irt}
\title{Fit an IRT model}
\usage{
fit_irt(irt_data, model = "rasch", iterations = 600, n_chains = 4)
}
\arguments{
\item{irt_data}{List formatted as in irt_data_bb()}

\item{model}{Type of IRT model to fit (rasch, 2pl, pcm)}

\item{iterations}{Number of MCMC iterations per chain}

\item{n_chains}{Number of MCMC chains to run}
}
\value{
stan object
}
\description{
Fit an IRT model
}
\examples{
\dontrun{
irt_data_bb(TestResponses, im_scored)
}
}
