% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-pcm.R
\name{fit_pcm}
\alias{fit_pcm}
\title{Fit PCM Model}
\usage{
fit_pcm(irt_data, iterations = 600, n_chains = 4)
}
\arguments{
\item{irt_data}{List formatted as in irt_data_bb()}

\item{iterations}{Number of MCMC iterations per chain}

\item{n_chains}{Number of MCMC chains to run}
}
\value{
stan object
}
\description{
Fit PCM Model
}
\examples{
\dontrun{
fit_pcm()
}
}
