% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-rasch.R
\name{fit_rasch}
\alias{fit_rasch}
\title{Fit Rasch Model}
\usage{
fit_rasch(irt_data, iterations = 600, n_chains = 4)
}
\arguments{
\item{irt_data}{List formatted as in irt_data_bb()}

\item{iterations}{Number of MCMC iterations per chain}

\item{n_chains}{Number of MCMC chains to run}
}
\value{
stan object
}
\description{
Fit Rasch Model
}
\examples{
\dontrun{
fit_rasch()
}
}
