% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{getmode}
\alias{getmode}
\title{Calculate the mode of a vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{vector to find mode of}
}
\value{
The mode of \code{v}
}
\description{
Calculate the mode of a vector
}
\examples{
getmode(c(1,1,1,0,0))
}
