% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{irt_data_bb}
\alias{irt_data_bb}
\title{Format the black box data to fit the IRT model}
\usage{
irt_data_bb(fbi_bb_data, scored_responses)
}
\arguments{
\item{fbi_bb_data}{Dataset of the form of TestResponses}

\item{scored_responses}{Scored responses (result of score_bb_data())}
}
\value{
A list in the format needed for fit_irt()
}
\description{
Format the black box data to fit the IRT model
}
\examples{
\dontrun{
irt_data_bb(TestResponses, im_scored)
}
}
