% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{item_mcmc_intervals}
\alias{item_mcmc_intervals}
\title{Calculate MCMC intervals for item estimates}
\usage{
item_mcmc_intervals(stan_model_output)
}
\arguments{
\item{stan_model_output}{Stan object from an IRT analysis}
}
\value{
Tibble of 50% and 95% posterior intervals for each b estimate
}
\description{
Calculate MCMC intervals for item estimates
}
\examples{
item_mcmc_intervals(im_model)
}
