% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{person_mcmc_intervals}
\alias{person_mcmc_intervals}
\title{Calculate MCMC intervals for participant estimates}
\usage{
person_mcmc_intervals(stan_model_output)
}
\arguments{
\item{stan_model_output}{Stan object from an IRT analysis}
}
\value{
Tibble of 50% and 95% posterior intervals for each theta estimate
}
\description{
Calculate MCMC intervals for participant estimates
}
\examples{
person_mcmc_intervals(im_model)
}
