% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_proficiency_posteriors}
\alias{plot_proficiency_posteriors}
\title{Plots the proficiency posteriors from IRT analysis}
\usage{
plot_proficiency_posteriors(irt_stan_output)
}
\arguments{
\item{irt_stan_output}{Stan object from IRT analysis}
}
\value{
A ggplot object
}
\description{
Plots the proficiency posteriors from IRT analysis
}
\examples{
\dontrun{
plot_proficiency_posteriors(im_model)
}
}
