% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\name{score_bb_data}
\alias{score_bb_data}
\title{Score the FBI black box data}
\usage{
score_bb_data(fbi_bb_data, scoring_method)
}
\arguments{
\item{fbi_bb_data}{Dataset of the form of TestResponses}

\item{scoring_method}{Which scoring method to use: inconclusive_mcar, inconclusive_incorrect, partial_credit, no_consensus_mcar, no_consensus_incorrect}
}
\value{
A vector of scored responses (1 row in fbi_bb_data = 1 entry)
}
\description{
Score the FBI black box data
}
\examples{
score_bb_data(TestResponses, "inconclusive_mcar")
}
