% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-data-functions.R
\docType{data}
\name{TestResponses}
\alias{TestResponses}
\title{Results from FBI "black box" study}
\format{A data frame with 17,121 rows and 8 variables
\describe{
  \item{Examiner_ID}{}
  \item{Pair_ID}{}
  \item{Mating}{}
  \item{Latent_Value}{}
  \item{Compare_Value}{}
  \item{Inconclusive_Reason}{If Compare_Value is Inconclusive, what is the reason given?}
  \item{Exclusion_Reason}{If Compare_Value is Exclusion, what is the reason given?}
  \item{Difficulty}{}
}}
\source{
\url{https://www.fbi.gov/services/laboratory/scientific-analysis/counterterrorism-forensic-science-research/black-box-study-results}
}
\usage{
TestResponses
}
\description{
Results from FBI "black box" study
}
\keyword{datasets}
