% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackboxstudyR-package.R
\docType{package}
\name{blackboxstudyR-package}
\alias{blackboxstudyR-package}
\alias{blackboxstudyR}
\title{The 'blackboxstudyR' package.}
\description{
Fits basic IRT models to black box data
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
